// src/components/SelectionModal/BettingOdds/BettingOddsHyperlink/types.ts
export type BettingOddsHyperlinkProps = {
    componentID: string;
    gameCode: string;
    league: string;
    source: string;
}


export type BaseHyperlinks = {
    espnCfbGameURL: string;
    cbsCfbGameURL: string;
    foxCfbURL: string;
    espnNflGameURL: string;
    cbsNflGameURL: string;
    foxNflURL: string;
}